This commit includes several files
constants.py contains a nashConstants object that contains all
the constants used in the library.

alg27.py is a translation of alg27.py---Hooke-Jeeves minimization
into python.

vectorin.py is a translation of the vectorin.pas file for
soliciting input

calceps.py is a translation of calceps.pas for calculating the
machine epsilon value