This commit is a complete draft version of alg27.py
This is a draft because while it appears to contain all the
executable lines of alg27.pas it's not clear that the Pascal
begin-end statements match with the indentions in Python.  That
must be cleared up.

Also, there it a "2" in line 72, " temp := 2.0*B[i]-X[i];" in the
original Pascal.  I don't know what purpose the 2 serves.